fracture: split payment business and user panels
