+++ frontend/src/lib/components/UserListTabImpl.svelte
@@ -1,0 +1,582 @@
+<script lang="ts">
+	import { get } from 'svelte/store';
+	import { users, serverMembers, currentUser, channels, createDM, getDMChannelIdForUser, socket, assignRole, removeUserRole, banUser, roleDefinitions } from '$lib/socket';
+	import { layoutStore } from '$lib/layoutStore';
+	import type { User } from '$lib/socket';
+	import ContextMenu from '$lib/components/context-menu/ContextMenu.svelte';
+	import type { ContextMenuItem } from '$lib/context-menu/types';
+	import { resolveUserDisplayColor } from '$lib/accessibility';
+	import { displayEnhancementSettingsStore } from '$lib/displayEnhancements';
+	import { rememberPeople } from '$lib/peopleTracker';
+	import {
+		MAX_LOCAL_NICKNAME_LENGTH,
+		clearLocalNicknameForUser,
+		getLocalNicknameForUser,
+		getUserIdentityKey,
+		localNicknamesStore,
+		setLocalNicknameForUser
+	} from '$lib/localNicknames';
+	import {
+		buildRolePriority,
+		buildRoleLabelMap,
+		getRoleLabel,
+		isCurrentUserEntry,
+		sortUsersList,
+		matchesSearch,
+		matchesPresenceFilter,
+		buildUserMenuItems,
+		queuePayment,
+		startDMCall,
+		type BuildMenuContext
+	} from './userListHelpers';
+
+	let contextMenuUser: User | null = null;
+	let contextMenuPosition = { x: 0, y: 0 };
+	let showContextMenu = false;
+	let friendSearchQuery = '';
+	let friendPresenceFilter: 'all' | 'active' | 'away' | 'busy' | 'offline' = 'all';
+	let friendSortMode: 'role' | 'name' | 'status' = 'role';
+	let offlineSectionExpanded = false;
+
+	$: rolePriority = buildRolePriority($roleDefinitions);
+	$: roleLabelMap = buildRoleLabelMap($roleDefinitions);
+	$: isEnhanced = $displayEnhancementSettingsStore.betterFriendListEnabled;
+	$: localNickEnabled = $displayEnhancementSettingsStore.localNicknamesEnabled;
+
+	function getLocalNickname(user: User): string {
+		if (!localNickEnabled) return '';
+		const key = getUserIdentityKey(user);
+		return key ? $localNicknamesStore[key] || '' : '';
+	}
+
+	function getDisplayName(user: User): string {
+		return getLocalNickname(user) || user.username;
+	}
+
+	function hasContextLocalNickname(): boolean {
+		if (!contextMenuUser) return false;
+		return Boolean(getLocalNickname(contextMenuUser));
+	}
+
+	function promptSetContextLocalNickname(): void {
+		if (!contextMenuUser) return;
+		const currentNickname = getLocalNicknameForUser(contextMenuUser);
+		const draft = window.prompt(`Set local nickname (max ${MAX_LOCAL_NICKNAME_LENGTH} characters)`, currentNickname || contextMenuUser.username);
+		if (draft === null) return;
+		setLocalNicknameForUser(contextMenuUser, draft);
+		closeContextMenu();
+	}
+
+	function clearContextLocalNickname(): void {
+		if (!contextMenuUser) return;
+		clearLocalNicknameForUser(contextMenuUser);
+		closeContextMenu();
+	}
+
+	$: onlineOtherUsers = sortUsersList(
+		$users.filter((user) => {
+			if (!matchesSearch(user, friendSearchQuery, isEnhanced, getDisplayName)) return false;
+			return matchesPresenceFilter(user, friendPresenceFilter, false, isEnhanced);
+		}),
+		friendSortMode,
+		rolePriority,
+		isEnhanced
+	);
+	$: rememberPeople($users);
+	$: rememberPeople($serverMembers);
+
+	$: groupedUsers = (() => {
+		const groups: Record<string, User[]> = {};
+		for (const user of onlineOtherUsers) {
+			const role = user.highestRole || 'member';
+			if (!groups[role]) groups[role] = [];
+			groups[role].push(user);
+		}
+		for (const role of Object.keys(groups)) {
+			groups[role] = sortUsersList(groups[role], friendSortMode, rolePriority, isEnhanced);
+		}
+		return groups;
+	})();
+
+	$: sortedRoles = Object.keys(groupedUsers).sort((a, b) => (rolePriority[b] || 0) - (rolePriority[a] || 0));
+
+	$: offlineUsers = (() => {
+		const onlineDbIds = new Set($users.map(u => u.dbUserId).filter(Boolean));
+		return $serverMembers
+			.filter(m => !onlineDbIds.has(m.dbUserId))
+			.filter((user) => matchesSearch(user, friendSearchQuery, isEnhanced, getDisplayName))
+			.filter((user) => matchesPresenceFilter(user, friendPresenceFilter, true, isEnhanced))
+			.sort((a, b) => a.username.localeCompare(b.username));
+	})();
+
+	function handleUserClick(user: User) {
+		if (isCurrentUserEntry(user, $currentUser)) {
+			layoutStore.openNotes();
+			return;
+		}
+		openDirectConversationWithUser(user);
+	}
+
+	function handleRightClick(event: MouseEvent, user: User) {
+		event.preventDefault();
+		contextMenuUser = user;
+		contextMenuPosition = { x: event.clientX, y: event.clientY };
+		showContextMenu = true;
+	}
+
+	function closeContextMenu() {
+		showContextMenu = false;
+		contextMenuUser = null;
+	}
+
+	function handleContextMessage() {
+		if (!contextMenuUser) return;
+		if (isCurrentUserEntry(contextMenuUser, $currentUser)) {
+			layoutStore.openNotes();
+			closeContextMenu();
+			return;
+		}
+		openDirectConversationWithUser(contextMenuUser);
+		closeContextMenu();
+	}
+
+	function openDirectConversationWithUser(user: User): void {
+		const self = get(currentUser);
+		if (!self || isCurrentUserEntry(user, $currentUser)) return;
+		const dmId = getDMChannelIdForUser(self, user);
+		const existingDM = get(channels).find((channel) => channel.id === dmId);
+		if (existingDM) {
+			layoutStore.openDM(dmId, user);
+			return;
+		}
+		createDM(user.id);
+		layoutStore.showDMsTab();
+		const unsubscribe = channels.subscribe((allChannels) => {
+			const newDM = allChannels.find((channel) => channel.id === dmId || (channel.type === 'dm' && channel.otherUser?.id === user.id));
+			if (!newDM) return;
+			layoutStore.openDM(newDM.id, user);
+			unsubscribe();
+		});
+	}
+
+	function handleContextRequestPayment(): void {
+		if (!contextMenuUser || isCurrentUserEntry(contextMenuUser, $currentUser) || !contextMenuUser.dbUserId) return;
+		queuePayment('payment_request', contextMenuUser);
+		openDirectConversationWithUser(contextMenuUser);
+		closeContextMenu();
+	}
+
+	function handleContextManualCash(): void {
+		if (!contextMenuUser || isCurrentUserEntry(contextMenuUser, $currentUser) || !contextMenuUser.dbUserId) return;
+		queuePayment('manual_cash', contextMenuUser);
+		openDirectConversationWithUser(contextMenuUser);
+		closeContextMenu();
+	}
+
+	async function handleContextVoiceCall() {
+		if (!contextMenuUser || !$socket || isCurrentUserEntry(contextMenuUser, $currentUser)) return;
+		closeContextMenu();
+		await startDMCall($socket, contextMenuUser, false);
+	}
+
+	async function handleContextVideoCall() {
+		if (!contextMenuUser || !$socket || isCurrentUserEntry(contextMenuUser, $currentUser)) return;
+		closeContextMenu();
+		await startDMCall($socket, contextMenuUser, true);
+	}
+
+	function handleBanContextUser(): void {
+		if (!contextMenuUser?.dbUserId) return;
+		const confirmBan = window.confirm(`Ban ${contextMenuUser.username}? They will lose access until manually re-enabled.`);
+		if (!confirmBan) return;
+		const reasonInput = window.prompt('Ban reason (optional):', '') || '';
+		banUser(contextMenuUser.dbUserId, reasonInput);
+		closeContextMenu();
+	}
+
+	function handleAssignContextRole(roleName: 'admin' | 'mod') {
+		if (!contextMenuUser?.dbUserId) return;
+		assignRole(contextMenuUser.dbUserId, roleName);
+		closeContextMenu();
+	}
+
+	function handleRemoveContextRole(roleName: 'admin' | 'mod') {
+		if (!contextMenuUser?.dbUserId) return;
+		removeUserRole(contextMenuUser.dbUserId, roleName);
+		closeContextMenu();
+	}
+
+	function handleResetContextUserToMember() {
+		if (!contextMenuUser?.dbUserId) return;
+		removeUserRole(contextMenuUser.dbUserId, 'admin');
+		removeUserRole(contextMenuUser.dbUserId, 'mod');
+		closeContextMenu();
+	}
+
+	function buildMenuCtx(): BuildMenuContext {
+		return {
+			contextMenuUser,
+			currentUser: $currentUser,
+			rolePriority,
+			localNicknamesEnabled: localNickEnabled,
+			hasLocalNickname: hasContextLocalNickname(),
+			socket: $socket
+		};
+	}
+
+	$: rawMenuItems = buildUserMenuItems(buildMenuCtx());
+	$: userMenuItems = rawMenuItems.map((item: ContextMenuItem) => {
+		const handlers: Record<string, () => void> = {
+			message: handleContextMessage,
+			'request-payment': handleContextRequestPayment,
+			'record-cash': handleContextManualCash,
+			voice: handleContextVoiceCall,
+			video: handleContextVideoCall,
+			'nickname-set': promptSetContextLocalNickname,
+			'nickname-clear': clearContextLocalNickname,
+			'make-admin': () => handleAssignContextRole('admin'),
+			'remove-admin': () => handleRemoveContextRole('admin'),
+			'make-mod': () => handleAssignContextRole('mod'),
+			'remove-mod': () => handleRemoveContextRole('mod'),
+			'reset-member': handleResetContextUserToMember,
+			'ban-user': handleBanContextUser
+		};
+		if (handlers[item.id]) return { ...item, onSelect: handlers[item.id] };
+		return item;
+	});
+
+	function getDisplayColor(user: User): string {
+		return resolveUserDisplayColor(user.roleColor, user.color);
+	}
+</script>
+
+<div class="user-list-tab">
+	{#if $displayEnhancementSettingsStore.betterFriendListEnabled}
+		<div class="friend-toolbar">
+			<input
+				type="text"
+				class="friend-search"
+				placeholder="Search people..."
+				bind:value={friendSearchQuery}
+			/>
+			<div class="friend-toolbar-row">
+				<select class="friend-select" bind:value={friendPresenceFilter}>
+					<option value="all">All</option>
+					<option value="active">Online</option>
+					<option value="away">Away</option>
+					<option value="busy">Busy</option>
+					<option value="offline">Offline</option>
+				</select>
+				<select class="friend-select" bind:value={friendSortMode}>
+					<option value="role">Role</option>
+					<option value="name">Name</option>
+					<option value="status">Status</option>
+				</select>
+			</div>
+		</div>
+	{/if}
+
+	{#each sortedRoles as role}
+		<div class="role-group">
+			<div class="role-header">
+				{getRoleLabel(role, roleLabelMap)} - {groupedUsers[role].length}
+			</div>
+			{#each groupedUsers[role] as user (user.id)}
+				<button
+					class="user-row"
+					on:click={() => handleUserClick(user)}
+					on:contextmenu={(e) => handleRightClick(e, user)}
+				>
+					<div class="user-avatar-wrap">
+						{#if user.profilePicture}
+							<img src={user.profilePicture} alt={getDisplayName(user)} class="user-avatar" />
+						{:else}
+							<div class="user-avatar-placeholder" style="--avatar-color: {user.color}">
+								{getDisplayName(user).charAt(0).toUpperCase()}
+							</div>
+						{/if}
+						<span class="presence-dot" class:active={user.status === 'active'} class:away={user.status === 'away'} class:busy={user.status === 'busy'}></span>
+					</div>
+					<div class="user-info">
+						<span class="user-display-name" style="color: {getDisplayColor(user)}">
+							{getDisplayName(user)}
+						</span>
+						{#if user.handle}
+							<span class="user-handle">@{user.handle}</span>
+						{/if}
+					</div>
+				</button>
+			{/each}
+		</div>
+	{/each}
+
+	{#if offlineUsers.length > 0}
+		<div class="role-group">
+			<button class="role-header offline-toggle" on:click={() => (offlineSectionExpanded = !offlineSectionExpanded)}>
+				<span class="offline-chevron" class:expanded={offlineSectionExpanded}>›</span>
+				Offline — {offlineUsers.length}
+			</button>
+			{#if offlineSectionExpanded}
+				{#each offlineUsers as user (user.dbUserId)}
+					<button
+						class="user-row offline"
+						on:click={() => handleUserClick(user)}
+						on:contextmenu={(e) => handleRightClick(e, user)}
+					>
+						<div class="user-avatar-wrap">
+							{#if user.profilePicture}
+								<img src={user.profilePicture} alt={getDisplayName(user)} class="user-avatar" />
+							{:else}
+								<div class="user-avatar-placeholder" style="--avatar-color: {user.color}">
+									{getDisplayName(user).charAt(0).toUpperCase()}
+								</div>
+							{/if}
+							<span class="presence-dot"></span>
+						</div>
+						<div class="user-info">
+							<span class="user-display-name" style="color: {getDisplayColor(user)}">
+								{getDisplayName(user)}
+							</span>
+							{#if user.handle}
+								<span class="user-handle">@{user.handle}</span>
+							{/if}
+						</div>
+					</button>
+				{/each}
+			{/if}
+		</div>
+	{/if}
+
+	{#if sortedRoles.length === 0 && offlineUsers.length === 0}
+		<div class="empty-state">No people match the current filters.</div>
+	{/if}
+
+	<ContextMenu
+		open={showContextMenu && !!contextMenuUser}
+		x={contextMenuPosition.x}
+		y={contextMenuPosition.y}
+		items={userMenuItems}
+		ariaLabel="User list actions"
+		headerLabel={contextMenuUser ? getDisplayName(contextMenuUser) : null}
+		on:close={closeContextMenu}
+	/>
+</div>
+
+<style>
+	.user-list-tab {
+		flex: 1;
+		overflow-y: auto;
+		padding: 0.4rem 0.35rem 0.7rem;
+		text-align: left;
+	}
+
+	.friend-toolbar {
+		display: flex;
+		flex-direction: column;
+		gap: 0.38rem;
+		padding: 0.35rem 0.45rem 0.55rem;
+		border-bottom: 1px solid color-mix(in srgb, var(--border-subtle) 70%, transparent);
+		margin-bottom: 0.3rem;
+	}
+
+	.friend-search,
+	.friend-select {
+		width: 100%;
+		height: 30px;
+		border-radius: 6px;
+		border: 1px solid var(--border-subtle);
+		background: var(--surface-raised);
+		color: var(--text-heading);
+		padding: 0 0.55rem;
+		font-size: 0.76rem;
+	}
+
+	.friend-toolbar-row {
+		display: grid;
+		grid-template-columns: minmax(0, 1fr) minmax(0, 1fr);
+		gap: 0.35rem;
+	}
+
+	.empty-state {
+		padding: 2rem 1rem;
+		text-align: center;
+		color: var(--text-secondary);
+		font-size: 0.875rem;
+	}
+
+	.role-group {
+		margin-bottom: 0.55rem;
+	}
+
+	.role-header {
+		padding: 0.45rem 0.3rem 0.3rem;
+		font-size: 0.7rem;
+		font-weight: 600;
+		text-transform: uppercase;
+		color: var(--text-secondary);
+		letter-spacing: 0.04em;
+		text-align: left;
+	}
+
+	.offline-toggle {
+		display: flex;
+		align-items: center;
+		gap: 0.35rem;
+		width: 100%;
+		background: none;
+		border: none;
+		cursor: pointer;
+		border-radius: 6px;
+		transition: color 0.15s, background 0.15s;
+		padding: 0.45rem 0.3rem 0.3rem;
+	}
+
+	.offline-toggle:hover {
+		color: var(--text-heading);
+		background: color-mix(in srgb, var(--accent-primary) 6%, transparent);
+	}
+
+	.offline-chevron {
+		font-size: 0.9rem;
+		line-height: 1;
+		display: inline-block;
+		transition: transform 0.18s ease;
+		transform: rotate(0deg);
+	}
+
+	.offline-chevron.expanded {
+		transform: rotate(90deg);
+	}
+
+	.user-row {
+		display: flex;
+		align-items: center;
+		justify-content: flex-start;
+		gap: 0.5rem;
+		padding: 0.48rem 0.35rem;
+		width: 100%;
+		background: transparent;
+		border: none;
+		color: var(--text-heading);
+		cursor: pointer;
+		text-align: left;
+		border-radius: 10px;
+		transition:
+			background 0.15s,
+			transform 0.15s;
+		margin-bottom: 0.08rem;
+	}
+
+	.user-row:hover {
+		background: color-mix(in srgb, var(--accent-primary) 8%, transparent);
+		transform: translateY(-1px);
+	}
+
+	.user-row.offline {
+		opacity: 0.58;
+	}
+
+	.user-avatar-wrap {
+		position: relative;
+		flex-shrink: 0;
+		width: 32px;
+		height: 32px;
+	}
+
+	.user-avatar,
+	.user-avatar-placeholder {
+		width: 32px;
+		height: 32px;
+		border-radius: 50%;
+		object-fit: cover;
+	}
+
+	.user-avatar-placeholder {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		font-size: 0.8rem;
+		font-weight: 600;
+		color: white;
+	}
+
+	.presence-dot {
+		position: absolute;
+		bottom: -2px;
+		right: -2px;
+		width: 14px;
+		height: 14px;
+		border-radius: 50%;
+		border: 3px solid var(--surface-base, #1a1a2e);
+		background: var(--status-offline, #6b7280);
+		box-sizing: border-box;
+		z-index: 2;
+		transition: transform 0.15s ease;
+	}
+
+	.presence-dot.active {
+		background: var(--status-online, #22c55e);
+		box-shadow: 0 0 0 0 rgba(var(--status-online-rgb, 34, 197, 94), 0.45);
+		animation: status-pulse-online 2.5s infinite ease-in-out;
+	}
+
+	.presence-dot.away { background: var(--status-away, #f59e0b); }
+	.presence-dot.busy { background: var(--status-busy, #ef4444); }
+
+	@keyframes status-pulse-online {
+		0%, 100% { box-shadow: 0 0 0 0 rgba(var(--status-online-rgb, 34, 197, 94), 0.45); }
+		50% { box-shadow: 0 0 0 5px rgba(var(--status-online-rgb, 34, 197, 94), 0); }
+	}
+
+	.user-info {
+		display: flex;
+		flex-direction: column;
+		min-width: 0;
+		gap: 1px;
+	}
+
+	.user-display-name {
+		font-size: 0.875rem;
+		font-weight: 600;
+		white-space: nowrap;
+		overflow: hidden;
+		text-overflow: ellipsis;
+	}
+
+	.user-handle {
+		font-size: 0.7rem;
+		color: var(--text-secondary);
+		white-space: nowrap;
+		overflow: hidden;
+		text-overflow: ellipsis;
+	}
+
+	@media (max-width: 768px) {
+		.friend-search,
+		.friend-select {
+			font-size: 16px;
+		}
+
+		.user-row {
+			padding: 0.625rem 0.75rem;
+			min-height: 52px;
+		}
+
+		.user-avatar-wrap {
+			width: 36px;
+			height: 36px;
+		}
+
+		.user-avatar,
+		.user-avatar-placeholder {
+			width: 36px;
+			height: 36px;
+		}
+
+		.user-display-name { font-size: 1rem; }
+		.user-handle { font-size: 0.8rem; }
+
+	}
+
+	.user-avatar-placeholder { background-color: var(--avatar-color, var(--accent-primary)); }
+</style>
